# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "MicroVibe-Go API",
    "description": "基于 AI 推荐算法的多端短视频平台后端 API，提供视频推荐、社交互动、直播等功能",
    "version": "1.0.0",
    "contact": {
      "name": "MicroVibe Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    },
    {
      "url": "https://api.microvibe.com",
      "description": "生产环境"
    }
  ],
...
  "paths": {
...
    "/api/v1/oauth/login": {
      "get": {
        "summary": "第三方登录",
        "tags": ["OAuth"],
        "parameters": [
          {
            "name": "provider",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": ["github", "google", "wechat"]
            },
            "description": "OAuth 提供商"
          }
        ],
        "responses": {
          "302": {
            "description": "重定向到 OAuth 提供商登录页面"
          }
        }
      }
    },
    "/api/v1/oauth/callback": {
      "get": {
        "summary": "OAuth 回调",
        "tags": ["OAuth"],
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "授权码"
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "状态参数"
          }
        ],
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Response"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "user": {
                              "$ref": "#/components/schemas/User"
                            },
                            "token": {
                              "type": "string",
                              "description": "JWT Token"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidParam"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "使用 JWT Token 进行身份验证，格式: Bearer {token}"
      }
    },
...
    "schemas": {
      "Response": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "业务状态码，0表示成功"
          },
          "message": {
            "type": "string",
            "description": "提示信息"
          },
          "data": {
            "description": "响应数据"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "uint"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "username": {
            "type": "string",
            "description": "用户名"
          },
          "email": {
            "type": "string",
            "description": "邮箱"
          },
          "phone": {
            "type": "string",
            "description": "手机号"
          },
          "nickname": {
            "type": "string",
            "description": "昵称"
          },
          "avatar": {
            "type": "string",
            "description": "头像URL"
          },
          "gender": {
            "type": "integer",
            "enum": [0, 1, 2],
            "description": "性别：0-未知，1-男，2-女"
          },
          "birthday": {
            "type": "string",
            "format": "date-time",
            "description": "生日"
          },
          "signature": {
            "type": "string",
            "description": "个性签名"
          },
          "province": {
            "type": "string",
            "description": "省份"
          },
          "city": {
            "type": "string",
            "description": "城市"
          },
          "follow_count": {
            "type": "integer",
            "format": "int64",
            "description": "关注数"
          },
          "follower_count": {
            "type": "integer",
            "format": "int64",
            "description": "粉丝数"
          },
          "video_count": {
            "type": "integer",
            "format": "int64",
            "description": "视频数"
          },
          "like_count": {
            "type": "integer",
            "format": "int64",
            "description": "获赞数"
          },
          "favorite_count": {
            "type": "integer",
            "format": "int64",
            "description": "收藏数"
          },
          "status": {
            "type": "integer",
            "enum": [0, 1],
            "description": "状态：0-禁用，1-正常"
          },
          "is_verified": {
            "type": "boolean",
            "description": "是否认证"
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "description": "最后登录时间"
          }
        }
      },
...
    },
    "responses": {
...
      "InvalidParam": {
        "description": "参数错误",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "example": 400
                },
                "message": {
                  "type": "string",
                  "example": "参数错误"
                },
                "data": {
                  "type": "object",
                  "nullable": true
                }
              }
            }
          }
        }
      },
...
    }
  }
}